% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tojson.R
\name{as.jsonld}
\alias{as.jsonld}
\title{Convert to JSON-LD}
\usage{
as.jsonld(input, ...)
}
\arguments{
\item{input}{Input object}

\item{...}{Further args passed on to \code{\link[jsonlite]{toJSON}}}
}
\description{
Convert to JSON-LD
}
\examples{
\dontrun{
out <- make_taxon(genus="Poa", epithet="annua", authority="L.")
x <- taxonref(rank="species", name="Homo sapiens", id=3454, uri="http://things.com")
as.jsonld(x)
}
}

