% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomen-package.R
\docType{package}
\name{binomen-package}
\alias{binomen}
\alias{binomen-package}
\title{Taxonomic class specification and parsing methods}
\description{
Taxonomic class specification and parsing methods
}
\keyword{package}

