% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomia-package.R
\docType{package}
\name{binomia-package}
\alias{binomia}
\alias{binomia-package}
\title{Taxonomic class specification and parsing methods}
\description{
Taxonomic class specification and parsing methods
}
\keyword{package}

