% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxon.R
\docType{class}
\name{binomial-class}
\alias{binomial-class}
\title{An S4 class to represent a taxonomic binomial}
\description{
An S4 class to represent a taxonomic binomial
}
\section{Slots}{

\describe{
\item{\code{genus}}{A genus name}

\item{\code{epithet}}{A specific epithet name}

\item{\code{canonical}}{Canonical name}

\item{\code{species}}{Species, genus plus epithet}

\item{\code{authority}}{Authority name}
}}

