% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{binomial}
\alias{binomial}
\title{A class to represent a taxonomic binomial}
\usage{
binomial(genus = NULL, epithet = NULL, canonical = NULL, species = NULL,
  authority = NULL)
}
\arguments{
\item{genus}{A genus name}

\item{epithet}{A specific epithet name}

\item{canonical}{Canonical name}

\item{species}{Species, genus plus epithet}

\item{authority}{Authority name}
}
\description{
A class to represent a taxonomic binomial
}
\examples{
binomial("Poa")
binomial("Poa", "annua", authority="L.")
}

