% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxon.R
\docType{class}
\name{classification-class}
\alias{classification-class}
\title{An S4 class to represent a taxonomic binomial}
\description{
An S4 class to represent a taxonomic binomial
}
\section{Slots}{

\describe{
\item{\code{kingdom}}{A kingdom name}

\item{\code{subkingdom}}{A subkingdom name}

\item{\code{infrakingdom}}{A infrakingdom name}

\item{\code{division}}{A division name}

\item{\code{phylum}}{A phylum name}

\item{\code{subdivision}}{A subdivision name}

\item{\code{infradavision}}{A infradavision name}

\item{\code{superclass}}{A superclass name}

\item{\code{clazz}}{A clazz name}

\item{\code{subclass}}{A subclass name}

\item{\code{infraclass}}{A infraclass name}

\item{\code{superorder}}{A superorder name}

\item{\code{order}}{A order name}

\item{\code{suborder}}{A suborder name}

\item{\code{infraorder}}{A infraorder name}

\item{\code{superfamily}}{A superfamily name}

\item{\code{family}}{A family name}

\item{\code{subfamily}}{A subfamily name}

\item{\code{tribe}}{A tribe name}

\item{\code{subtribe}}{A subtribe name}

\item{\code{genus}}{A genus name}

\item{\code{subgenus}}{A subgenus name}

\item{\code{section}}{A section name}

\item{\code{subsection}}{A subsection name}

\item{\code{species}}{A species name}

\item{\code{subspecies}}{A subspecies name}

\item{\code{variety}}{A variety name}

\item{\code{race}}{A race name}

\item{\code{subvariety}}{A subvariety name}

\item{\code{stirp}}{A stirp name}

\item{\code{morph}}{A morph name}

\item{\code{form}}{A form name}

\item{\code{aberration}}{A aberration name}

\item{\code{subform}}{A subform name}

\item{\code{unspecified}}{A unspecified name}
}}

