% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxon.R
\docType{class}
\name{taxonref-class}
\alias{taxonref-class}
\title{An S4 class to represent a taxonomic reference}
\description{
An S4 class to represent a taxonomic reference
}
\section{Slots}{

\describe{
\item{\code{rank}}{Taxonomic rank}

\item{\code{name}}{A name}

\item{\code{id}}{Identifier}

\item{\code{source}}{Source of name}
}}

