% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gethier.R
\docType{methods}
\name{gethier}
\alias{gethier}
\alias{gethier,taxon-method}
\title{Get hierarchy as a data.frame}
\usage{
gethier(x)

\S4method{gethier}{taxon}(x)
}
\arguments{
\item{data}{An object of class taxon}
}
\description{
Get hierarchy as a data.frame
}
\examples{
\dontrun{
(out <- taxon(genus="Poa", epithet="annua", authority="L.",
             family='Poaceae', clazz='Poales', kingdom='Plantae', variety='annua'))
# get hierarchy as data.frame
gethier(out)
}
}

