% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{grouping}
\alias{grouping}
\title{A class to represent a taxonomic classification}
\usage{
grouping(kingdom = NULL, subkingdom = NULL, infrakingdom = NULL,
  division = NULL, phylum = NULL, subdivision = NULL,
  infradavision = NULL, superclass = NULL, clazz = NULL,
  subclass = NULL, infraclass = NULL, superorder = NULL, order = NULL,
  suborder = NULL, infraorder = NULL, superfamily = NULL, family = NULL,
  subfamily = NULL, tribe = NULL, subtribe = NULL, genus = NULL,
  subgenus = NULL, section = NULL, subsection = NULL, species = NULL,
  subspecies = NULL, variety = NULL, race = NULL, subvariety = NULL,
  stirp = NULL, morph = NULL, form = NULL, aberration = NULL,
  subform = NULL, unspecified = NULL)
}
\arguments{
\item{kingdom}{A kingdom name}

\item{subkingdom}{A subkingdom name}

\item{infrakingdom}{A infrakingdom name}

\item{division}{A division name}

\item{phylum}{A phylum name}

\item{subdivision}{A subdivision name}

\item{infradavision}{A infradavision name}

\item{superclass}{A superclass name}

\item{clazz}{A clazz name}

\item{subclass}{A subclass name}

\item{infraclass}{A infraclass name}

\item{superorder}{A superorder name}

\item{order}{A order name}

\item{suborder}{A suborder name}

\item{infraorder}{A infraorder name}

\item{superfamily}{A superfamily name}

\item{family}{A family name}

\item{subfamily}{A subfamily name}

\item{tribe}{A tribe name}

\item{subtribe}{A subtribe name}

\item{genus}{A genus name}

\item{subgenus}{A subgenus name}

\item{section}{A section name}

\item{subsection}{A subsection name}

\item{species}{A species name}

\item{subspecies}{A subspecies name}

\item{variety}{A variety name}

\item{race}{A race name}

\item{subvariety}{A subvariety name}

\item{stirp}{A stirp name}

\item{morph}{A morph name}

\item{form}{A form name}

\item{aberration}{A aberration name}

\item{subform}{A subform name}

\item{unspecified}{A unspecified name}
}
\description{
A class to represent a taxonomic classification
}
\examples{
grouping(kingdom=taxonref("kingdom", "Animalia"),
               species=taxonref("species", "Homo sapiens"))
}

