% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods.R
\docType{class}
\name{ListOfTaxa-class}
\alias{ListOfTaxa-class}
\title{An S4 class to represent a list of taxa}
\description{
An S4 class to represent a list of taxa
}
\section{Slots}{

\describe{
\item{\code{taxon}}{An object of class taxon}
}}

