% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxon.R
\docType{class}
\name{ListOfTaxonRefs-class}
\alias{ListOfTaxonRefs-class}
\title{An S4 class to represent a list of taxonomic names}
\description{
An S4 class to represent a list of taxonomic names
}
\section{Slots}{

\describe{
\item{\code{names}}{Names}
}}

