% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_taxon.R
\name{make_taxon}
\alias{make_taxon}
\title{Make taxon from class}
\usage{
make_taxon(genus = "none", epithet = "none", authority = "none", ...)
}
\arguments{
\item{genus}{Genus name, e.g., Homo (in Homo sapiens)}

\item{epithet}{Specific epithet, e.g., sapiens (in Homo sapiens)}

\item{authority}{Taxonomic authority}

\item{...}{Further args.}
}
\description{
Make taxon from class
}
\examples{
\dontrun{
(out <- make_taxon(genus="Poa"))
(out <- make_taxon(genus="Poa", epithet="annua", authority="L."))
(out <- make_taxon(genus="Poa", epithet="annua", authority="L.",
                   family='Poaceae', clazz='Poales', kingdom='Plantae', variety='annua'))
out$binomial
out$binomial$canonical
out$binomial$species
out$binomial$authority
out$classification
out$classification$family
out \%>\% pick(family) # get a single rank
out \%>\% span(kingdom, family) # get a range of ranks
gethier(out) # get hierarchy as data.frame
}
}

