% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_taxon_fromclass.R
\name{make_taxon_fromclass}
\alias{make_taxon_fromclass}
\title{Make taxon from class}
\usage{
make_taxon_fromclass(data, authority = "none")
}
\arguments{
\item{data}{A data.frame}

\item{authority}{Taxonomic authority}
}
\description{
Make taxon from class
}
\examples{
df <- data.frame(rank=c('family','tribe','subtribe','genus','subgenus','species'),
                 name=c('Helianthi','Helianthi','Helianthi','Poa','Festuci','Poa annua'),
                 id=c(1,2,3,4,5,6),
                 stringsAsFactors = FALSE)
apply(df, 1, make_taxon_fromclass)
}

