% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parsing.R
\name{parsing}
\alias{name}
\alias{parsing}
\alias{range}
\alias{select}
\alias{uri}
\title{Select, parse, etc. taxanomic names}
\usage{
select(.data, ...)

range(.data, ...)

name(.data)

uri(.data)
}
\arguments{
\item{.data}{Input, object of class taxon}

\item{...}{Further unnamed args, see examples}
}
\description{
Select, parse, etc. taxanomic names
}
\examples{
out <- make_taxon(genus="Poa", epithet="annua", authority="L.",
   family='Poaceae', clazz='Poales', kingdom='Plantae', variety='annua')
# get single name
out \%>\% select(family)
out \%>\% select(genus)
out \%>\% select(species)
out \%>\% select(species) \%>\% name()
out \%>\% select(species) \%>\% uri()
# get range of names
out \%>\% range(kingdom, genus)
}

