% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{name}
\alias{name.taxonref}
\alias{pick}
\alias{uri}
\alias{uri.taxonref}
\title{Pick names}
\usage{
pick(.data, ...)

name(.data)

\method{name}{taxonref}(.data)

uri(.data)

\method{uri}{taxonref}(.data)
}
\arguments{
\item{.data}{Input, object of class taxon}

\item{...}{Further unnamed args, see examples}
}
\description{
Pick names
}
\examples{
# operating on `taxon` objects
out <- make_taxon(genus="Poa", epithet="annua", authority="L.",
   family='Poaceae', clazz='Poales', kingdom='Plantae', variety='annua')
# get single name
out \%>\% pick(family)
out \%>\% pick(genus)
out \%>\% pick(species)
out \%>\% pick(species) \%>\% name()
out \%>\% pick(species) \%>\% uri()

# operating on taxonomic data.frames
df <- data.frame(class=c('Magnoliopsida','Magnoliopsida','Magnoliopsida',
                         'Magnoliopsida','Magnoliopsida','Magnoliopsida'),
         order=c('Asterales','Asterales','Fagales','Poales','Poales','Poales'),
         family=c('Asteraceae','Asteraceae','Fagaceae','Poaceae','Poaceae','Poaceae'),
         genus=c('Helianthus','Helianthus','Quercus','Poa','Festuca','Holodiscus'),
         stringsAsFactors = FALSE)
(df2 <- taxon_df(df))

## select single taxonomic class
df2 \%>\% pick(order, Fagales)
df2 \%>\% pick(family, Asteraceae)
df2 \%>\% pick(genus, Poa)
}

