% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\title{Pop names out, drop them that is}
\usage{
pop(.data, ...)
}
\arguments{
\item{.data}{Input, object of class taxon}

\item{...}{Further unnamed args, see examples}
}
\value{
For \code{taxon} inputs gives back a \code{taxon} object. For \code{taxondf}
inputs, gives back a \code{taxondf} object.
}
\description{
Pop names out, drop them that is
}
\examples{
# operating on `taxon` objects
out <- make_taxon(genus="Poa", epithet="annua", authority="L.",
   family='Poaceae', clazz='Poales', kingdom='Plantae', variety='annua')
## single names
out \%>\% pop(family)
out \%>\% pop(genus)
out \%>\% pop(species)
## many names
out \%>\% pop(family, genus, species)

# operating on taxonomic data.frames
df <- data.frame(class=c('Magnoliopsida','Magnoliopsida','Magnoliopsida',
                         'Magnoliopsida','Magnoliopsida','Magnoliopsida'),
         order=c('Asterales','Asterales','Fagales','Poales','Poales','Poales'),
         family=c('Asteraceae','Asteraceae','Fagaceae','Poaceae','Poaceae','Poaceae'),
         genus=c('Helianthus','Helianthus','Quercus','Poa','Festuca','Holodiscus'),
         stringsAsFactors = FALSE)
(df2 <- taxon_df(df))

## pop out a single taxonomic class
df2 \%>\% pop(order)
df2 \%>\% pop(family)
df2 \%>\% pop(genus)
}

