% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\alias{range.taxon}
\alias{range.taxondf}
\title{Parse taxon or taxondf objects by a range of names}
\usage{
range(.data, ...)

\method{range}{taxon}(.data, ...)

\method{range}{taxondf}(.data, ...)
}
\arguments{
\item{.data}{Input, object of class taxon}

\item{...}{Further unnamed args, see examples}
}
\description{
Parse taxon or taxondf objects by a range of names
}
\examples{
# operating on `taxon` objects
out <- make_taxon(genus="Poa", epithet="annua", authority="L.",
   family='Poaceae', clazz='Poales', kingdom='Plantae', variety='annua')
out \%>\% range(kingdom, genus)

# operating on taxonomic data.frames
df <- data.frame(class=c('Magnoliopsida','Magnoliopsida','Magnoliopsida',
                         'Magnoliopsida','Magnoliopsida','Magnoliopsida'),
         order=c('Asterales','Asterales','Fagales','Poales','Poales','Poales'),
         family=c('Asteraceae','Asteraceae','Fagaceae','Poaceae','Poaceae','Poaceae'),
         genus=c('Helianthus','Helianthus','Quercus','Poa','Festuca','Holodiscus'),
         stringsAsFactors = FALSE)
(df2 <- taxon_df(df))

## filter to get a range of classes
df2 \%>\% range(order, genus)
df2 \%>\% range(family, genus)
}

