% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomen-package.R
\docType{data}
\name{rank_table}
\alias{rank_table}
\title{Lookup-table for IDs of taxonomic ranks}
\value{
data.frame with two columns and 34 rows. The two columns:
\itemize{
 \item rankid integer; smaller numbers are higher ranks.
 \item ranks character; rank name. Some rows have more than one name,
 in which the names are considered of equal height.
}
}
\description{
Lookup-table for IDs of taxonomic ranks
}
\keyword{data}
