% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatter each taxon in a taxondf to a taxon object}
\usage{
scatter(x, ...)
}
\arguments{
\item{x}{A taxonomic data.frame}

\item{...}{Further args, ignored for now}
}
\description{
Scatter each taxon in a taxondf to a taxon object
}
\examples{
# operating on taxonomic data.frames
df <- data.frame(class=c('Magnoliopsida','Magnoliopsida','Magnoliopsida',
                         'Magnoliopsida','Magnoliopsida','Magnoliopsida'),
         order=c('Asterales','Asterales','Fagales','Poales','Poales','Poales'),
         family=c('Asteraceae','Asteraceae','Fagaceae','Poaceae','Poaceae','Poaceae'),
         genus=c('Helianthus','Helianthus','Quercus','Poa','Festuca','Holodiscus'),
         stringsAsFactors = FALSE)
(df2 <- taxon_df(df))

## scatter each taxon into a taxon class
df2 \%>\% scatter(order)
}

