% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxon-package.R
\docType{package}
\name{taxon-package}
\alias{taxon}
\alias{taxon-package}
\title{Taxonomic information from around the web.}
\description{
Taxnomic search and phylogeny retrieval.
}
\keyword{package}

