% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{taxon}
\alias{taxon}
\title{A class to represent a single taxon}
\usage{
taxon(binomial, grouping)
}
\arguments{
\item{binomial}{A binomial name}

\item{grouping}{A grouping object}
}
\description{
A class to represent a single taxon
}
\examples{
bin <- binomial("Poa", "annua", authority="L.")
class <- grouping(kingdom=taxonref("kingdom", "Plantae"),
   species=taxonref("family", "Poaceae"))
taxon(bin, class)

# many names input
splist <- c('Litsea bindoniana', 'Rubus ghanakantae', 'Desmanthus palmeri',
 'Leptinella longipes', 'Asarum sakawanum', 'Cistanche compacta',
 'Ormosia nanningensis', 'Claoxylon physocarpum', 'Hedycarya arborea',
 'Hypnum gracile')
lapply(splist, binomial)
}
