% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{taxon_classes}
\alias{taxon_classes}
\title{Taxonomic class methods}
\description{
Taxonomic class methods
}
\details{
The taxonomic classes:

\itemize{
 \item \strong{binomial} - A binomial, via \code{\link{binomial}}
 \item \strong{taxonref} - A single taxonref, via \code{\link{taxonref}}
 \item \strong{taxonrefs} - A list of taxonrefs, via \code{\link{taxonrefs}}
 \item \strong{grouping} - A grouping (classification) object, via \code{\link{grouping}}
 \item \strong{taxon} - A taxon, via \code{\link{taxon}}
 \item \strong{taxa} - List of taxon objects, via \code{\link{taxa}}
}
}
