% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_df.R
\name{taxon_df}
\alias{taxon_df}
\title{Taxon data.frame}
\usage{
taxon_df(x)
}
\arguments{
\item{x}{A data.frame of taxa}
}
\description{
Taxon data.frame
}
\examples{
# subset data.frame using taxonomy
df <- data.frame(family=c('Asteraceae','Asteraceae','Asteraceae','Poaceae','Poaceae','Poaceae'),
                 tribe=c('Helianthi','Helianthi','Helianthi','Poaeae','Festuci','Poaeae'),
                 genus=c('Helianthus','Helianthus','Madia','Poa','Festuca','Holodiscus'),
                 stringsAsFactors = FALSE)
df2 <- taxon_df(df)
df2 \%>\% pick(family)
df2 \%>\% pick(genus, tribe)
}

