% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3methods.R
\name{taxonref}
\alias{taxonref}
\title{A class to represent a taxonomic reference}
\usage{
taxonref(rank = "none", name = "none", id = "none", uri = "none")
}
\arguments{
\item{rank}{(character) Taxonomic rank}

\item{name}{(character) A name}

\item{id}{(character,numeric) Identifier}

\item{uri}{(character) Source of name}
}
\description{
A class to represent a taxonomic reference
}
\examples{
taxonref("genus", "Poa", 56, "http://scottchamberlain.info/")
}

