% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{taxonrefs}
\alias{taxonrefs}
\title{A class to represent a list of taxonomic references}
\usage{
taxonrefs(...)
}
\arguments{
\item{...}{One or more taxonref objects}
}
\description{
A class to represent a list of taxonomic references
}
\examples{
a <- taxonref("genus", "Poa", 56, "http://scottchamberlain.info/")
b <- taxonref("genus", "Quercus", 32343, "http://scottchamberlain.info/")
taxonrefs(a, b)
}
