% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange}
\alias{arrange}
\title{Arrange taxon or taxondf objects by one or more choices}
\usage{
arrange(.data, ...)
}
\arguments{
\item{.data}{Input, object of class taxon}

\item{...}{Further unnamed args, see examples}
}
\description{
Arrange taxon or taxondf objects by one or more choices
}
\examples{
# operating on taxonomic data.frames
df <- data.frame(class=c('Magnoliopsida','Magnoliopsida','Magnoliopsida',
                         'Magnoliopsida','Magnoliopsida','Magnoliopsida'),
         order=c('Asterales','Asterales','Fagales','Poales','Poales','Poales'),
         family=c('Asteraceae','Asteraceae','Fagaceae','Poaceae','Poaceae','Poaceae'),
         genus=c('Helianthus','Helianthus','Quercus','Poa','Festuca','Holodiscus'),
         stringsAsFactors = FALSE)
(df2 <- taxon_df(df))

## arrange the taxonomic data.frame
df2 \%>\% arrange(order)
df2 \%>\% arrange(family)
df2 \%>\% arrange(genus)
df2 \%>\% arrange(genus, order)
}

