% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethier.R
\name{gethier}
\alias{gethier}
\title{Get hierarchy as a data.frame}
\usage{
gethier(x)
}
\arguments{
\item{x}{An object of class taxon}
}
\description{
Get hierarchy as a data.frame
}
\examples{
\dontrun{
bin <- binomial("Poa", "annua", authority="L.")
class <- grouping(kingdom=taxonref("kingdom", "Plantae"),
   species=taxonref("family", "Poaceae"))
out <- taxon(bin, class)
# get hierarchy as data.frame
gethier(out)
}
}
