% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxon.R
\docType{class}
\name{ListOfTaxa-class}
\alias{ListOfTaxa-class}
\title{An S4 class to represent a taxonomic binomial}
\description{
An S4 class to represent a taxonomic binomial
}
\section{Slots}{

\describe{
\item{\code{taxon}}{An object of class taxon}
}}

