% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parsing.R
\name{parsing}
\alias{parsing}
\title{Select, parse, etc. taxanomic names}
\description{
Select, parse, etc. taxanomic names
}
\examples{
# operating on `taxon` objects
out <- make_taxon(genus="Poa", epithet="annua", authority="L.",
   family='Poaceae', clazz='Poales', kingdom='Plantae', variety='annua')
# get single name
out \%>\% select(family)
out \%>\% select(genus)
out \%>\% select(species)
out \%>\% select(species) \%>\% name()
out \%>\% select(species) \%>\% uri()
# get range of names
out \%>\% span(kingdom, genus)

# operating on taxonomic data.frames
df <- data.frame(class=c('Magnoliopsida','Magnoliopsida','Magnoliopsida',
                         'Magnoliopsida','Magnoliopsida','Magnoliopsida'),
         order=c('Asterales','Asterales','Fagales','Poales','Poales','Poales'),
         family=c('Asteraceae','Asteraceae','Fagaceae','Poaceae','Poaceae','Poaceae'),
         genus=c('Helianthus','Helianthus','Quercus','Poa','Festuca','Holodiscus'),
         stringsAsFactors = FALSE)
(df2 <- taxon_df(df))

## select single taxonomic class
df2 \%>\% select(order, Fagales)
df2 \%>\% select(family, Asteraceae)
df2 \%>\% select(genus, Poa)

## filter to get a range of classes
df2 \%>\% span(order, genus)
df2 \%>\% span(family, genus)

## combine them
 df2 \%>\%
   span(family, genus) \%>\%
   select(family, Asteraceae)
}

