% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parts.R
\name{parts}
\alias{id}
\alias{id.taxon}
\alias{id.taxonref}
\alias{name}
\alias{name.taxon}
\alias{name.taxonref}
\alias{parts}
\alias{rank}
\alias{rank.taxon}
\alias{rank.taxonref}
\alias{uri}
\alias{uri.taxon}
\alias{uri.taxonref}
\title{Pick names}
\usage{
name(.data)

\method{name}{taxon}(.data)

\method{name}{taxonref}(.data)

uri(.data)

\method{uri}{taxon}(.data)

\method{uri}{taxonref}(.data)

rank(.data)

\method{rank}{taxon}(.data)

\method{rank}{taxonref}(.data)

id(.data)

\method{id}{taxon}(.data)

\method{id}{taxonref}(.data)
}
\arguments{
\item{.data}{Input, object of class taxon}
}
\value{
For \code{taxon} inputs, gives back a \code{taxonref} object. For \code{taxondf}
inputs, gives back \code{taxondf}.
}
\description{
Pick names
}
\examples{
# operating on `taxon` objects
out <- make_taxon(genus="Poa", epithet="annua", authority="L.",
   family='Poaceae', clazz='Poales', kingdom='Plantae', variety='annua')

out \%>\% name()
out \%>\% uri()
out \%>\% rank()
out \%>\% id()
}

