% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sort.R
\name{sort}
\alias{desc}
\alias{sort}
\alias{sort_}
\alias{sort_.taxondf}
\title{Sort taxon or taxondf objects by one or more choices}
\usage{
sort(x, ...)

sort_(x, ..., .dots)

\method{sort_}{taxondf}(x, ..., .dots)

desc(x)
}
\arguments{
\item{x}{Input, object of class taxondf}

\item{...}{Comma separated list of unquoted expressions. You can treat variable names
like they are positions. Use positive values to select variables; use negative values
to drop variables.}

\item{.dots}{Use sort_() to do standard evaluation}
}
\description{
Sort taxon or taxondf objects by one or more choices
}
\examples{
# operating on taxonomic data.frames
df <- data.frame(class=c('Magnoliopsida','Magnoliopsida','Magnoliopsida',
                         'Magnoliopsida','Magnoliopsida','Magnoliopsida'),
         order=c('Asterales','Asterales','Fagales','Poales','Poales','Poales'),
         family=c('Asteraceae','Asteraceae','Fagaceae','Poaceae','Poaceae','Poaceae'),
         genus=c('Helianthus','Helianthus','Quercus','Poa','Festuca','Holodiscus'),
         stringsAsFactors = FALSE)
(df2 <- taxon_df(df))

## sort the taxonomic data.frame
df2 \%>\% sort(order)
df2 \%>\% sort(family)
df2 \%>\% sort(genus)
df2 \%>\% sort(genus, order)

## reverse order
df2 \%>\% sort(desc(genus))
}

