% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3methods.R
\name{taxa}
\alias{taxa}
\title{An S4 class to represent a list of taxa}
\usage{
taxa(...)
}
\arguments{
\item{...}{An object of class taxon}
}
\description{
An S4 class to represent a list of taxa
}
\examples{
bin <- binomial("Poa", "annua", authority="L.")
class <- classification(kingdom=taxonref("kingdom", "Plantae"),
   species=taxonref("family", "Poaceae"))
taxa(taxon(bin, class), taxon(bin, class))
}

