% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxon.R
\docType{class}
\name{taxon-class}
\alias{taxon-class}
\title{An S4 class to represent a taxonomic binomial}
\description{
An S4 class to represent a taxonomic binomial
}
\section{Slots}{

\describe{
\item{\code{binomial}}{A binomial name}

\item{\code{classification}}{A classification object}
}}

