% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxon.R
\name{taxon}
\alias{taxon}
\title{Make taxon from class}
\usage{
taxon(genus = "none", epithet = "none", authority = "none", ...)
}
\arguments{
\item{genus}{Genus name, e.g., Homo (in Homo sapiens)}

\item{epithet}{Specific epithet, e.g., sapiens (in Homo sapiens)}

\item{authority}{Taxonomic authority}

\item{...}{Further args.}
}
\description{
Make taxon from class
}
\examples{
\dontrun{
(out <- taxon(genus="Poa", epithet="annua", authority="L."))
(out <- taxon(genus="Poa", epithet="annua", authority="L.",
                   family='Poaceae', clazz='Poales', kingdom='Plantae', variety='annua'))
(out <- taxon(genus="Poa"))
out@binomial
out@binomial@canonical
out@binomial@species
out@binomial@authority
out@classification
out@classification@family
out[['family']] # get a single rank
out['family','kingdom'] # get a range of ranks
out['variety','genus'] # get a range of ranks
gethier(out) # get hierarchy as data.frame
}
}

