% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods.R
\name{taxon_classes}
\alias{taxon_classes}
\title{Taxonomic S4 class methods}
\description{
Taxonomic S4 class methods
}
\details{
The taxonomic classes:

\itemize{
 \item binomial A binomial
 \item taxonref A single taxonref
 \item ListOfTaxonRefs A list of taxonrefs
 \item classification A classification object
 \item taxon A taxon
 \item ListOfTaxa List of taxon objects
}
}

