% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3methods.R
\name{taxon_classes}
\alias{taxon_classes}
\title{Taxonomic S4 class methods}
\description{
Taxonomic S4 class methods
}
\details{
The taxonomic classes:

\itemize{
 \item binomial A binomial, via \code{\link{binomial}}
 \item taxonref A single taxonref, via \code{\link{taxonref}}
 \item taxonrefs A list of taxonrefs, via \code{\link{taxonrefs}}
 \item classification A classification object, via \code{\link{classification}}
 \item taxon A taxon, via \code{\link{taxon}}
 \item taxa List of taxon objects, via \code{\link{taxa}}
}
}

