% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxon_df.R
\name{taxon_df}
\alias{taxon_df}
\title{Taxon data.frame}
\usage{
taxon_df(data)
}
\arguments{
\item{data}{A data.frame of taxa}
}
\description{
Taxon data.frame
}
\examples{
\dontrun{
# subset data.frame using taxonomy
df <- data.frame(family=c('Asteraceae','Asteraceae','Asteraceae','Poaceae','Poaceae','Poaceae'),
                 tribe=c('Helianthi','Helianthi','Helianthi','Poaeae','Festuci','Poaeae'),
                 genus=c('Helianthus','Helianthus','Madia','Poa','Festuca','Holodiscus'),
                 stringsAsFactors = FALSE)
df2 <- taxon_df(df)
df2['family','Asteraceae']
df2['genus','Madia']
df2['tribe','Helianthi']
}
}

